Spellbook Launch
Converted most of the functionality to jQuery UI so that we can use
their themes, and had a bit more ease of programming.

Also implemented snap.js which is the basis for the spellbook.

Spellbook is fully implemented. Right now it slides out from the right,
allows you to drag and drop spells in there, and persists between
browser refreshes. You can reorder the spells and the ordering persists
too.